types: Generate types according to architecture

+++ kernel/include/multiboot.h
@@ -128,10 +128,10 @@ void multiboot_dump_info    (multiboot_info_t *info);
 
 #define FOREACH_MEMORY_MAP(MMAP, INFO) \
   for(multiboot_memory_map_t *MMAP = \
-        (multiboot_memory_map_t *)(unsigned long)INFO->mmap_addr; \
-      (unsigned long)MMAP < INFO->mmap_addr + INFO->mmap_len; \
+        (multiboot_memory_map_t *)(uintptr_t)INFO->mmap_addr; \
+      (uintptr_t)MMAP < INFO->mmap_addr + INFO->mmap_len; \
       MMAP = (multiboot_memory_map_t *) \
-        ((unsigned long)MMAP + MMAP->size + sizeof(MMAP->size)) \
+        ((uintptr_t)MMAP + MMAP->size + sizeof(MMAP->size)) \
       )
 
 #endif /* end of include guard: MULTIBOOT_H_RKNBOMGQ */
